i learned how to create a function
i did not learn this in w3schools i did a bit of my own research and i watched someone break down how to create functions and call them and now i have a much better understanding of what i am learning when functions are involved. This has helped a lot and i will continue to do this.